Move HostingResolved out of firebaseConfig
